Update README.md
Project Process: A fast contraband detection system was designed using PMMW images for object detection. The project goals were defined, and challenges in resolution, contrast, and noise in PMMW images were analyzed. The PMMW image dataset was collected and preprocessed. YOLOv8 was chosen as the base model, and structural optimizations were made. GSConv was used in the Backbone to reduce redundant features, and GSConv-Slim was introduced in the Neck to maintain accuracy while reducing complexity. LADH was added to the detection head for efficient, accurate, and lightweight detection. The model was trained and evaluated for optimal performance.

Project Results: The model size was 2.3M with a compression rate of 26.98%, enabling lightweight deployment. This significantly improved safety detection efficiency in public places like airports and subway stations, effectively enhancing public safety.